# Use ES 2015 syntax to generate more compact code for blocking functions.