The idea is that we can avoid referencing each variable identifier multiple times in blocking functions intro and outro using ES2015 syntax sugar for object literals and descructuring assignments.

Consider following example:

```go
package main

var c = make(chan string, 1)

func producer() {
	message := "hi"
	c <- message
}

func consumer() string {
	return <-c
}

func main() {
	go producer()
	println(consumer())
}
```

Generated code of the `producer()` function before this change:

```js
	producer = function() {
		var message, $s, $r;
		/* */ $s = 0; var $f, $c = false; if (this !== undefined && this.$blk !== undefined) { $f = this; $c = true; message = $f.message; $s = $f.$s; $r = $f.$r; } s: while (true) { switch ($s) { case 0:
		message = "hi";
		$r = $send(c, message); /* */ $s = 1; case 1: if($c) { $c = false; $r = $r.$blk(); } if ($r && $r.$blk !== undefined) { break s; }
		$s = -1; return;
		/* */ } return; } if ($f === undefined) { $f = { $blk: producer }; } $f.message = message; $f.$s = $s; $f.$r = $r; return $f;
	};
```

We can see that the local variable `message` is referenced 3 times in lines 1-2 and twice on the last line.

Generated code after this change:

```js
	producer = function() {
		var {message, $s, $r, $c, $f} = $restore(this, {});
		/* */ $s = $s || 0; s: while (true) { switch ($s) { case 0:
		message = "hi";
		$r = $send(c, message); /* */ $s = 1; case 1: if($c) { $c = false; $r = $r.$blk(); } if ($r && $r.$blk !== undefined) { break s; }
		$s = -1; return;
		/* */ } return; } if ($f === undefined) { $f = { $blk: producer }; } $f = {...$f, $r, message, $s};return $f;
	};
```

Local variable `message` is referenced only once on the line 1 and once more on the last line, plus overall the function intro code is a lot more compact.

Reference app: [jQuery TodoMVC](https://github.com/gopherjs/todomvc) (`acf500a6c32a83d8c4582d967b09a65febf0e120`)

|          BRANCH          |            ORIGINAL             |            MINIFIED             |       COMPRESSED (GZIP)        |
|--------------------------|---------------------------------|---------------------------------|--------------------------------|
| Pull request (es6-loads) | 2,904,199 bytes                 | 1,913,886 bytes                 | 381,037 bytes                  |
| Target branch (master)   | 8.83% decrease (-281,161 bytes) | 7.86% decrease (-163,201 bytes) | 3.23% decrease (-12,722 bytes) |

Updates #136.