@nevkontakte 2022-08-07
Good question, I assumed it should be more or less neutral, but the GoroutineSwitching benchmark shows a significant degradation:

```
$ benchstat /tmp/old.txt /tmp/new.txt
name                old time/op  new time/op  delta
GoroutineSwitching   356ns ± 3%  1218ns ± 2%  +241.96%  (p=0.000 n=10+8)
```

In real life it probably isn't so bad — our CI run time did not change much, but I'll look into that and see if we can improve it.

At some future point I was also considering reimplementing goroutines using ES 2015 generator functions, which should be even more compact and hopefully performant, but that would be a much bigger change.