@nevkontakte 2022-08-13
After some investigation I was able to make this change more or less performance-neutral, and even push size reduction a bit further.

Here are some benchmarking results:

```
Env: Node.js 12 / WSL
name                old time/op  new time/op  delta
GoroutineSwitching   350ns ± 3%   343ns ± 3%  -1.84%  (p=0.004 n=10+9)
Env: Chrome 104 / Win10
name                old time/op  new time/op  delta
GoroutineSwitching   266ns ± 2%   253ns ± 4%  -4.85%  (p=0.000 n=9+10)
Env: Firefox 103 / Win10
name                old time/op  new time/op  delta
GoroutineSwitching   684ns ±14%   735ns ±29%   ~     (p=0.156 n=9+10)
Env: Edge 104 / Win 10
name                old time/op  new time/op  delta
GoroutineSwitching   251ns ± 2%   268ns ± 1%  +6.77%  (p=0.000 n=9+10)
```

A couple of caveats for the benchmark numbers:

 - Performance deltas for Node, Chrome and Edge are likely explained by variable background load on my PC. I didn't bother setting up clean benchmarking environment, so I would consider this neutral overall.
 - Firefox seems to show some decrease in performance, but it also shows very high variability in the performance numbers, to the point that benchstat calls the difference statistically insignificant. Unfortunately, its profiling tools lack the ability of per-line profile, so I was unable to pinpoint the difference.

|          BRANCH          |            ORIGINAL             |            MINIFIED             |       COMPRESSED (GZIP)        |
|--------------------------|---------------------------------|---------------------------------|--------------------------------|
| Pull request (es6-loads) | 2,877,935 bytes                 | 1,895,863 bytes                 | 380,638 bytes                  |
| Target branch (master)   | 9.65% decrease (-307,425 bytes) | 8.72% decrease (-181,224 bytes) | 3.33% decrease (-13,121 bytes) |