@nevkontakte 2022-08-13
Overall it's  worth noting that the like with every synthetic benchmark, the real-world difference is likely to be much, much smaller. `GoroutineSwitching` benchmark was designed to cause goroutines block on each other as aggressively as possible, which is unlikely to happen in real-life code.