@nevkontakte 2022-08-13
As a curious side note, gopherjs on Node 12 performs in this benchmark only two times slower than native go:

```
name                old time/op  new time/op  delta
GoroutineSwitching   157ns ± 0%   343ns ± 3%  +118.22%  (p=0.000 n=9+9)
```

Hat tip to V8 devs for that.