@paralin 2022-08-16
@nevkontakte thanks for your work on this and gopherjs in general. I can try to test this PR on a fairly complex project and see if it works. Unfortunately some packages I'm importing started using generics... Will try to see if I can help implement that perhaps.