@nevkontakte 2022-08-16
Thanks @paralin, any additional testing is always welcome, although I'm fairly confident in this PR and was going to merge it as soon as I address a couple of minor comments.

If you'd like to help adding support for generics, it will be much appreciated. I've been doing some research on how it could be done in GopherJS context and will try to write that down https://github.com/gopherjs/gopherjs/issues/1013 today after work.