@tomconnell-wf 2022-08-16
I'm a fan of this.  Our app compiles to 5.6MB minified-and-gzipped, so even a 3% decrease in size is going to be at least 100-some kb smaller, which is significant.